App page improvements
- safe old dialog code correctly in dialogs/
- reduce code duplication in PackagePage and SnapPage headers with AppHeaderData
- SnapPage: hide connections if not installed
