+++ lib/store_app/common/app_header.dart
@@ -16,14 +16,121 @@
  */
 
 import 'package:flutter/material.dart';
-import 'package:software/l10n/l10n.dart';
-import 'package:url_launcher/url_launcher.dart';
-import 'package:yaru_colors/yaru_colors.dart';
-import 'package:yaru_icons/yaru_icons.dart';
+import 'package:software/store_app/common/app_infos.dart';
+import 'package:software/store_app/common/app_website.dart';
 
 const headerStyle = TextStyle(fontWeight: FontWeight.w500, fontSize: 14);
 
-class AppHeader extends StatelessWidget {
+class OneColumnAppHeader extends StatelessWidget {
+  const OneColumnAppHeader({
+    super.key,
+    required this.headerData,
+  });
+
+  final AppHeaderData headerData;
+
+  @override
+  Widget build(BuildContext context) {
+    final screenWidth = MediaQuery.of(context).size.width;
+    final wideEnough = screenWidth > 750;
+    final height = wideEnough ? 150.0 : 40.0;
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        SizedBox(
+          height: height,
+          child: headerData.icon,
+        ),
+        SizedBox(width: wideEnough ? 30 : 10),
+        Expanded(
+          child: SizedBox(
+            height: height,
+            child: Column(
+              crossAxisAlignment: wideEnough
+                  ? CrossAxisAlignment.start
+                  : CrossAxisAlignment.end,
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    if (wideEnough)
+                      Text(
+                        headerData.title,
+                        style: Theme.of(context).textTheme.headline3,
+                        overflow: TextOverflow.ellipsis,
+                      ),
+                    if (wideEnough)
+                      AppWebsite(
+                        website: headerData.website,
+                        verified: headerData.verified,
+                        publisherName: headerData.publisherName,
+                      ),
+                  ],
+                ),
+                headerData.controls,
+              ],
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+class TwoColumnAppHeader extends StatelessWidget {
+  const TwoColumnAppHeader({
+    super.key,
+    required this.properties,
+  });
+
+  final AppHeaderData properties;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.center,
+      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        Column(
+          children: [
+            SizedBox(
+              height: 180,
+              child: properties.icon,
+            ),
+            Text(
+              properties.title,
+              style: Theme.of(context).textTheme.headline3,
+              overflow: TextOverflow.ellipsis,
+              textAlign: TextAlign.center,
+            ),
+            AppWebsite(
+              website: properties.website,
+              verified: properties.verified,
+              publisherName: properties.publisherName,
+            ),
+          ],
+        ),
+        properties.controls,
+        AppInfos(
+          strict: properties.strict,
+          confinementName: properties.confinementName,
+          license: properties.license,
+          installDate: properties.installDate,
+          installDateIsoNorm: properties.installDateIsoNorm,
+          version: properties.version,
+        ),
+        Text(
+          properties.summary,
+          style: Theme.of(context).textTheme.bodySmall,
+        ),
+      ],
+    );
+  }
+}
+
+class AppHeaderData {
   final Widget? icon;
   final String title;
   final String summary;
@@ -36,193 +143,21 @@ class AppHeader extends StatelessWidget {
   final bool verified;
   final String publisherName;
   final String website;
+  final Widget controls;
 
-  const AppHeader({
-    super.key,
-    required this.icon,
+  AppHeaderData({
+    this.icon,
     required this.title,
     required this.summary,
-    required this.version,
     required this.strict,
     required this.confinementName,
+    required this.version,
     required this.license,
     required this.installDate,
     required this.installDateIsoNorm,
     required this.verified,
     required this.publisherName,
     required this.website,
+    required this.controls,
   });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            SizedBox(
-              height: 65,
-              child: icon,
-            ),
-            const SizedBox(
-              width: 15,
-            ),
-            SizedBox(
-              width: 300,
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Expanded(
-                        child: Row(
-                          children: [
-                            Expanded(
-                              child: Text(
-                                title,
-                                overflow: TextOverflow.visible,
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      const SizedBox(
-                        width: 10,
-                      ),
-                      // if (snapIsInstalled)
-                    ],
-                  ),
-                  const SizedBox(
-                    height: 5,
-                  ),
-                  Text(
-                    summary,
-                    style: Theme.of(context).textTheme.bodySmall,
-                    overflow: TextOverflow.ellipsis,
-                    maxLines: 10,
-                  ),
-                  SizedBox(
-                    height: 30,
-                    child: Row(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        if (verified)
-                          Tooltip(
-                            message: context.l10n.verified,
-                            child: const Icon(
-                              Icons.verified,
-                              size: 20,
-                              color: YaruColors.success,
-                            ),
-                          ),
-                        if (website.isNotEmpty)
-                          InkWell(
-                            borderRadius: BorderRadius.circular(8),
-                            onTap: () => launchUrl(Uri.parse(website)),
-                            child: Row(
-                              children: [
-                                Padding(
-                                  padding: EdgeInsets.only(
-                                    left: verified ? 5 : 0,
-                                    right: 5,
-                                  ),
-                                  child: Text(
-                                    publisherName,
-                                    style: const TextStyle(fontSize: 15),
-                                  ),
-                                ),
-                                Padding(
-                                  padding:
-                                      EdgeInsets.only(right: verified ? 5 : 0),
-                                  child: Icon(
-                                    YaruIcons.external_link,
-                                    color:
-                                        Theme.of(context).colorScheme.onSurface,
-                                    size: 18,
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                      ],
-                    ),
-                  )
-                ],
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(
-          height: 20,
-        ),
-        Row(
-          children: [
-            Column(
-              children: [
-                Text(
-                  context.l10n.confinement,
-                  style: headerStyle,
-                ),
-                Row(
-                  children: [
-                    Icon(
-                      strict ? YaruIcons.shield : YaruIcons.warning,
-                      size: 18,
-                    ),
-                    const SizedBox(
-                      width: 5,
-                    ),
-                    Text(
-                      confinementName,
-                      style: Theme.of(context).textTheme.bodyMedium,
-                    ),
-                  ],
-                ),
-              ],
-            ),
-            const SizedBox(height: 50, width: 30, child: VerticalDivider()),
-            Column(
-              mainAxisSize: MainAxisSize.min,
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Flexible(
-                  child: Text(context.l10n.license, style: headerStyle),
-                ),
-                Flexible(
-                  child: Tooltip(
-                    message: license,
-                    child: Text(
-                      license.split(' ').first,
-                      overflow: TextOverflow.ellipsis,
-                      style: Theme.of(context).textTheme.bodyMedium,
-                    ),
-                  ),
-                )
-              ],
-            ),
-            const SizedBox(height: 50, width: 30, child: VerticalDivider()),
-            Column(
-              children: [
-                Text(
-                  installDate.isEmpty
-                      ? context.l10n.version
-                      : context.l10n.installDate,
-                  style: headerStyle,
-                ),
-                Tooltip(
-                  message: installDateIsoNorm,
-                  child: Text(
-                    installDate.isEmpty ? version : installDate,
-                    style: headerStyle.copyWith(fontWeight: FontWeight.normal),
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ],
-    );
-  }
 }
