+++ lib/store_app/common/snap_page.dart
@@ -25,12 +25,11 @@ import 'package:software/store_app/common/app_infos.dart';
 import 'package:software/store_app/common/app_media.dart';
 import 'package:software/store_app/common/border_container.dart';
 import 'package:software/store_app/common/constants.dart';
-import 'package:software/store_app/common/one_column_app_header.dart';
+import 'package:software/store_app/common/app_header.dart';
 import 'package:software/store_app/common/page_layouts.dart';
 import 'package:software/store_app/common/snap_connections_settings.dart';
 import 'package:software/store_app/common/snap_controls.dart';
 import 'package:software/store_app/common/snap_model.dart';
-import 'package:software/store_app/common/two_column_app_header.dart';
 import 'package:ubuntu_service/ubuntu_service.dart';
 import 'package:yaru_icons/yaru_icons.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
@@ -88,40 +87,43 @@ class _SnapPageState extends State<SnapPage> {
         ),
         child: AppDescription(description: model.description ?? ''),
       ),
-      BorderContainer(
-        padding: const EdgeInsets.only(
-          bottom: pagePadding,
-          right: pagePadding,
-        ),
-        child: SnapConnectionsSettings(connections: model.connections),
-      )
+      if (model.snapIsInstalled)
+        BorderContainer(
+          padding: const EdgeInsets.only(
+            bottom: pagePadding,
+            right: pagePadding,
+          ),
+          child: SnapConnectionsSettings(connections: model.connections),
+        )
     ];
+
+    final headerData = AppHeaderData(
+      confinementName: model.confinement != null ? model.confinement!.name : '',
+      icon: InkWell(
+        borderRadius: BorderRadius.circular(100),
+        onTap: model.installDate.isNotEmpty ? model.open : null,
+        child: YaruSafeImage(
+          url: model.iconUrl,
+          fallBackIconData: YaruIcons.snapcraft,
+        ),
+      ),
+      installDate: model.installDate,
+      installDateIsoNorm: model.installDateIsoNorm,
+      license: model.license ?? '',
+      strict: model.strict,
+      verified: model.verified,
+      publisherName: model.publisher?.displayName ?? '',
+      website: model.storeUrl ?? '',
+      summary: model.summary ?? '',
+      title: model.title ?? '',
+      version: model.version,
+      controls: const SnapControls(),
+    );
 
     final oneColumnAppHeader = BorderContainer(
       padding: const EdgeInsets.all(pagePadding),
-      // width: 500,
       child: OneColumnAppHeader(
-        confinementName:
-            model.confinement != null ? model.confinement!.name : '',
-        icon: InkWell(
-          borderRadius: BorderRadius.circular(100),
-          onTap: model.installDate.isNotEmpty ? model.open : null,
-          child: YaruSafeImage(
-            url: model.iconUrl,
-            fallBackIconData: YaruIcons.snapcraft,
-          ),
-        ),
-        installDate: model.installDate,
-        installDateIsoNorm: model.installDateIsoNorm,
-        license: model.license ?? '',
-        strict: model.strict,
-        verified: model.verified,
-        publisherName: model.publisher?.displayName ?? '',
-        website: model.storeUrl ?? '',
-        summary: model.summary ?? '',
-        title: model.title ?? '',
-        version: model.version,
-        controls: const SnapControls(),
+        headerData: headerData,
       ),
     );
 
@@ -129,25 +131,7 @@ class _SnapPageState extends State<SnapPage> {
       padding: const EdgeInsets.all(pagePadding),
       width: 500,
       child: TwoColumnAppHeader(
-        confinementName:
-            model.confinement != null ? model.confinement!.name : '',
-        icon: YaruSafeImage(
-          url: model.iconUrl,
-          fallBackIconData: YaruIcons.snapcraft,
-        ),
-        installDate: model.installDate,
-        installDateIsoNorm: model.installDateIsoNorm,
-        license: model.license ?? '',
-        strict: model.strict,
-        verified: model.verified,
-        publisherName: model.publisher?.displayName ?? '',
-        website: model.storeUrl ?? '',
-        summary: model.summary ?? '',
-        title: model.title ?? '',
-        version: model.version,
-        controls: const SnapControls(
-          direction: Axis.vertical,
-        ),
+        properties: headerData,
       ),
     );
